Extract RunListener implementation from InstrumentedTestCollector
Extracted RunListener implementation to a new class InstrumentedTestListener from InstrumentedTestCollector for better separation of concerns and testability.
